backend/requirements.txt
fastapi
uvicorn
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
pillow
python-multipart
pydantic
httpx
